Se mejoro el manejo de errores de MateriasController

+++ controllers/materiasController.js
@@ -22,8 +22,8 @@ const fetchMateria = async (req, res) => {
 
     res.json({ materia })
   } catch (err) {
-    console.log(err)
-    res.sendStatus(400)
+    console.error('(fetchMateria) Error al obtener la materia:', err)
+    res.status(500).json({ message: 'Error interno del servidor' })
   }
 }
 
@@ -35,6 +35,16 @@ const createMateria = async (req, res) => {
       year
     } = req.body
 
+    if (!nombre || !descripcion || !year) {
+      return res.status(400).json({ error: 'Falta un campo' })
+    }
+
+    const materiaExistente = await Materia.findOne({ nombre })
+
+    if (materiaExistente) {
+      res.status(409).json({ error: 'Ya existe un alumno con ese DNI' })
+    }
+
     const materia = await Materia.create({
       nombre,
       descripcion,
@@ -43,8 +53,8 @@ const createMateria = async (req, res) => {
 
     res.json({ materia })
   } catch (err) {
-    console.log(err)
-    res.sendStatus(400)
+    console.log('(createMateria) Error al crear alumno', err)
+    res.status(500).json({ error: 'Error interno del servidor' })
   }
 }
 
@@ -58,14 +68,19 @@ const updateMateria = async (req, res) => {
       year
     } = req.body
 
+    const materiaExistente = await Materia.findById(id)
+    if (!materiaExistente) {
+      return res.status(404).json({ error: 'No se ha encontrado el materia' })
+    }
+
     await Materia.findByIdAndUpdate(id, { nombre, descripcion, year })
 
     const materia = await Materia.findById(id)
 
     res.json({ materia })
   } catch (err) {
-    console.log(err)
-    res.sendStatus(400)
+    console.log('(updateMateria) Error al actualizar la materia:', err)
+    res.status(500).json({ error: 'Error interno del servidor' })
   }
 }
 
@@ -75,9 +90,19 @@ const deleteMateria = async (req, res) => {
 
     const materiaDel = await Materia.findById(id)
 
+    if (!materiaDel) {
+      return res.status(404).json({ error: 'No se ha encontrado la materia' })
+    }
+
     if (materiaDel.curso) {
       const curso = await Curso.findById(materiaDel.curso)
+
+      if (!curso) {
+        return res.status(404).json({ error: 'No se ha encontrado el curso' })
+      }
+
       curso.materias.pop({ _id: materiaDel._id })
+
       await curso.save()
     }
 
@@ -85,22 +110,29 @@ const deleteMateria = async (req, res) => {
 
     res.json({ success: `Se ha eliminado la materia ${materia.nombre}` })
   } catch (err) {
-    console.log(err)
-    res.sendStatus(400)
+    console.log('(deleteMateria) Error al eliminar la materia', err)
+    res.status(500).json({ error: 'Error interno del servidor' })
   }
 }
 
 const deleteMaterias = async (req, res) => {
   try {
     const _ids = req.body._ids
-    console.log(_ids)
+
+    if (!Array.isArray(_ids) || _ids.length === 0) {
+      return res.status(400).json({ error: 'Debe proporcionar un array de IDs valido' })
+    }
 
     const materias = await Materia.deleteMany({ _id: { $in: _ids } })
 
+    if (materias.deletedCount === 0) {
+      res.status(404).json({ error: 'No se encontraron materias para eliminar' })
+    }
+
     res.json({ success: `Se ha eliminado ${materias}` })
   } catch (err) {
-    console.log(err)
-    res.sendStatus(400)
+    console.log('(deleteMaterias) Error al eliminar las materis', err)
+    res.status(500).json({ error: 'Error interno del servidor' })
   }
 }
